requirements.txt
selenium==4.12.0
webdriver-manager==4.0.0
pandas==2.0.3
beautifulsoup4==4.12.2
flask==2.3.3
scikit-learn==1.3.0
joblib==1.3.1
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
google-search-results
python-dotenv==1.0.0
